==> demo_notebooks/Nb0_clean_WHOGMdb.ipynb <==
## Extract relevant information from WHOGMDatabase.csv - from Shaddick et al.

* Find ground monitor locations which have reported PM25 values later than 2015
* Get latest year
* create categorical country code
* add binary classification "is_urban"

Creates two data files:

* `world_pm25.csv`
* `north_america_pm25.csv`


import json
import pandas as pd
import numpy as np


who_clean = pd.read_csv('WHOGMDatabase.csv', encoding = "ISO-8859-1")

who_clean.keys()

who_clean['GBDSuperRegion'].unique()

who_clean['SDG2Region'].unique()

world = who_clean.drop(columns=['StationIDOrig',
    'StationIDOldDatabase', 'City', 'CityGiulia', 'CityClean',
    'PM25PercCoverage', 'PM25Grading',
    'PM10', 'PM10PercCoverage', 'PM10Grading',
    'LocationInfo', 'Source', 'MonitorTypeOrig', 'PM25Conv',                           
    'UnspecifiedType', 'WebLink', 'Version', 'WHOStatus', 'WHORegion',
    'WHOIncomeRegion', 'SDG1Region', 'SDG3Region',
    'GBDRegion', 'GBDSuperRegion'], axis= 1)

world.keys()

print(world.shape)
world = world[world['Year']>2015]
print(world.shape)

world = world.dropna(subset =['PM25'])
print(world.shape)

world = world.groupby(['StationID']).apply(lambda x: x.loc[x['Year'].idxmax()], include_groups=False).reset_index(drop=True)
print(world.shape)

world['is_urban'] = np.where(world['MonitorType'].isin(['Urban', 'urban', 'Industrial', 'industrial']), 1, 0)
world['country'] = world['ISO3'].astype('category').cat.codes + 1
world['log_PM25'] = np.log(world['PM25'])

world.head(10)

import csv
world.to_csv('world_pm25.csv', 
             columns=['Year', 'CityReverseGeocoded', 'Longitude', 'Latitude', 
                      'log_PM25', 'is_urban', 'country'], 
                     quoting=csv.QUOTE_NONNUMERIC, index=False)

euro_slice = world[world['SDG2Region'] == 'Europe']
euro_slice.shape

euro = euro_slice.copy(deep=True).reset_index(drop=True)
euro['country'] = euro['ISO3'].astype('category').cat.codes + 1
euro

euro.to_csv('euro_log_pm25.csv', 
             columns=['Year', 'CountryName', 'CityReverseGeocoded', 'Longitude', 'Latitude',
                      'log_PM25', 'is_urban', 'country'],
                     quoting=csv.QUOTE_NONNUMERIC, index=False)

na_slice = world[world['ISO3'].isin(['CAN','USA','MEX'])]
north_america = na_slice.copy(deep=True).reset_index(drop=True)
north_america['country'] = north_america['ISO3'].astype('category').cat.codes + 1

north_america.head(3)

north_america.tail(3)

print(north_america.shape)

north_america.to_csv('north_america_log_pm25.csv', 
             columns=['Year', 'CityReverseGeocoded', 'Longitude', 'Latitude', 
                      'log_PM25', 'is_urban', 'country'], 
                     quoting=csv.QUOTE_NONNUMERIC, index=False)
