==> demo_notebooks/Nb1_preprocess_coords.ipynb <==
## Step 1: Preprocess geo-locations - geo_spde.preprocess_coords

Given a geo-located dataset, identify unique geo-locations and project them to a coordinate system which respects the distance between points.   (footnote, citation needed).

#### Libraries

import os
import sys
from pathlib import Path
import numpy as np
import pandas as pd
import geopandas as gpd
import geodatasets
from plotnine import *

from geo_spde.coords import preprocess_coords
from geo_spde.exceptions import CoordsError

#### Geolocated data

Taken from Shaddick et al 2016:  Ground monitor measurements from Canada, USA, Mexico

log_pm25 = pd.read_csv('north_america_log_pm25.csv')
log_pm25.head(5)


#### World map

From naturalearth,  low res world map with country boundaries.

url = "https://naciscdn.org/naturalearth/110m/cultural/ne_110m_admin_0_countries.zip"
world = gpd.read_file(url)
sorted(world.keys())

na_map = world[world['ISO_A3'].isin(['CAN', 'USA', 'MEX'])]
na_map[['ISO_A3', 'ADMIN']]

#### Plot raw data

Show ground monitor measurements on map, raw coordinate system (lon/lat)

p1 = (ggplot() +
         geom_map(na_map, fill='white', color='black', size=0.2) +
         geom_point(data=log_pm25,
                   mapping=aes(x='Longitude', y='Latitude', color='log_PM25'),
                   size=0.1) +
         scale_color_gradient2(low='blue', mid='lightgreen', high='darkorange',
                               midpoint=1.5,  name='PM2.5') +
         theme_minimal() +
         labs(title="North America log_PM2.5 Locations, raw coords") +
         coord_fixed())

p1

### geo_spde.preprocess_coords

Remove any duplicate observations, project lon/lat to appropriate map scale.

log_pm25_mainland = log_pm25[log_pm25['Longitude'] > -130].reset_index(drop=True)
log_pm25_mainland = log_pm25_mainland[log_pm25_mainland['Longitude'] < -65].reset_index(drop=True)
log_pm25_coords = log_pm25_mainland[['Longitude','Latitude']].to_numpy()
clean_coords, indices, proj_info = preprocess_coords(log_pm25_coords)

### Plot preprocessed data

# Create dataframe with projected coordinates
df_projected = log_pm25_mainland.iloc[indices].copy()
df_projected['x_proj'] = clean_coords[:, 0]
df_projected['y_proj'] = clean_coords[:, 1]
df_projected.head(3)

na_map_albers = na_map.to_crs(proj_info['proj4_string'])

p2 = (ggplot() +
      geom_map(na_map_albers, fill='white', color='black', size=0.2) +
      geom_point(data=df_projected,
                 mapping=aes(x='x_proj*1000', y='y_proj*1000', color='log_PM25'),
                 size=0.1) +
         scale_color_gradient2(low='blue', mid='lightgreen', high='darkorange',
                               midpoint=1.5,  name='log_PM25') +
         theme_minimal() +
         labs(title="North America log PM2.5 Locations, Albers projection") +
         coord_fixed())

p2

p1
